api and client enhancements

+++ API/DTOs/QuoteRequestDto.cs
@@ -8,7 +8,7 @@ namespace API.DTOs
     public class QuoteRequestDto
     {
         public int UserId { get; set; }
-        public int LoanAmount { get; set; }
+        public double LoanAmount { get; set; }
         public int NumberOfRepayments { get; set; }
         public int ProductId { get; set; }
     }
